refactor: move evolution/promotion_gate.rs into astra-evolution
Move promotion_gate.rs (664 LOC) to astra-evolution::promotion_gate.
This module provides typed promotion verdicts for evolution proposals.

Changes:
- Move promotion_gate.rs with fixed imports (crate::types, crate::promotion_signals)
- Replace runtime stub with re-export

Depends on: PR #127 (promotion_signals extraction)
