+++ rust/crates/astra-evolution/src/lib.rs
@@ -6,6 +6,7 @@
 
 pub mod evolver;
 pub mod persistence;
+pub mod promotion_gate;
 pub mod promotion_signals;
 pub mod signal_collector;
 pub mod store;
